Migrate update nillable string array tests to new location
Also updates them to the new framework
